requirements.txt
pandas
numpy
matplotlib
seaborn
scikit-learn
faiss-cpu
streamlit
altair>=4.2,<5
